fixed folders and centered some pages such as forgotpasswod otpverify and reset password
