+++ client/src/Page/Admin/Tenant.jsx
@@ -3,7 +3,7 @@ import TenantCard from '../../Component/TenantCard'
 import SearchBar from '../../Component/SearchBar'
 import { FaPlus } from 'react-icons/fa6'
 import { useState } from 'react'
-import AddTenantForm from '../../Component/AddTenantForm'
+import AddTenantForm from '../../Component/AdminComponent/AddTenantForm'
 import { useDispatch, useSelector } from 'react-redux'
 import { createTenant, fetchUsers, handleSearchUser } from '../../features/user'
 import { fetchApartments } from '../../features/apartment'
