+++ client/src/Page/ForgotPass.jsx
@@ -1,9 +1,7 @@
 import React, { useState } from 'react'
 import logo from '/logo.svg'
-
 import ForgotPass1 from '/ForgotPass1.svg'
-import { useNavigate } from 'react-router-dom'
-import { base_url } from '../utils/constants'
+import { useNavigate, Link } from 'react-router-dom'
 const ForgotPass = () => {
   const [email, setEmail] = useState('')
   const [error, setError] = useState(null)
@@ -13,7 +11,7 @@ const ForgotPass = () => {
     try {
       setError(null)
       const response = await fetch(
-        `${base_url}/api/user/forgot-password?email=${email}`,
+        `${import.meta.env.VITE_URL}/api/user/forgot-password?email=${email}`,
         {
           method: 'POST',
         },
@@ -32,12 +30,14 @@ const ForgotPass = () => {
   return (
     <>
       <div className=" w-full h-screen py-10 px-10">
-        <img
-          src={logo}
-          alt="PinaupaPH logo"
-          className="lg:block lg:ml-10 hidden  "
-        />
-        <div className="lg:flex-row lg:ml-20  flex flex-col items-center">
+        <Link to="/">
+          <img
+            src={logo}
+            alt="PinaupaPH logo"
+            className="lg:block lg:ml-10 hidden  "
+          />
+        </Link>
+        <div className="lg:flex-row  flex flex-col items-center justify-center ">
           <div className="  lg:w-1/2 lg:mt-0 mt-5">
             <img
               src={logo}
@@ -53,7 +53,7 @@ const ForgotPass = () => {
 
           <div className="lg:shadow-md lg:rounded-md lg:shadow-dark-gray p-10">
             <div className="">
-              <h1 className="lg:text-3xl text-2xl font-bold text-primary">
+              <h1 className="lg:text-3xl text-2xl font-bold text-primary-color">
                 Password Recovery
               </h1>
               <p className="text-sm mt-3  text-dark-gray">
@@ -62,7 +62,10 @@ const ForgotPass = () => {
             </div>
 
             <form onSubmit={handleSubmit} className="  w-full mt-3 ">
-              <label htmlFor="email" className="text-primary font-bold text-lg">
+              <label
+                htmlFor="email"
+                className="text-primary-color font-bold text-lg"
+              >
                 Email
               </label>
               <input
@@ -72,7 +75,7 @@ const ForgotPass = () => {
                 placeholder="e.g juan.delacruz@gmail.com"
                 className="mr-10 rounded-md py-3 border-2 px-3 border-dark-gray w-full mt-2"
               />
-              <button className="lg:mt-10 lg:mb-4 bg-primary text-white w-full mt-5 py-3 px-2 rounded-md hover:opacity-80">
+              <button className="lg:mt-10 lg:mb-4 bg-primary-color text-white w-full mt-5 py-3 px-2 rounded-md hover:opacity-80">
                 Send
               </button>
             </form>
