+++ client/src/Page/Tenant/ProfileTenant.jsx
@@ -11,7 +11,7 @@ import EditFamMemTable from '../../Component/EditFamMemTable'
 //import { payment_url } from '../../utils/constants'
 import DocumentCard from '../../Component/DocumentCard'
 import AddHousehold from '../../Component/AddHousehold'
-import EditApartment from '../../Component/EditApartment'
+import EditApartment from '../../Component/AdminComponent/EditApartment'
 import TransactionTable from '../../Component/TransactionTable'
 import TransactionMobile from '../../Component/TransactionMobile'
 import { GrFormView, GrFormAdd } from 'react-icons/gr'
