+++ client/src/features/cctv.js
@@ -1,8 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
-import { base_url } from '../utils/constants'
 import Cookies from 'js-cookie'
 
-const token = Cookies.get('token')
 const cctvSlice = createSlice({
   name: 'cctv',
   initialState: {
