+++ client/src/features/report.js
@@ -1,5 +1,4 @@
 import { createSlice } from '@reduxjs/toolkit'
-import { base_url } from '../utils/constants'
 import Cookies from 'js-cookie'
 const reportSlice = createSlice({
   name: 'report',
