--- client/src/utils/constants.js
@@ -1,1 +1,0 @@
-export const base_url = 'http://localhost:5000'
